Refactor onewire options flow tests
